+++ CHANGELOG.md
@@ -9,6 +9,9 @@ All notable changes to this project should be documented in this file.
 - A `bump_version.py` script to automate version update, by @smiyashaji.
 - A `--timeout` CLI option for configurable script running timeouts, by @ShrayJP.
 
+### Enhanced
+
+- Timeout log files will be compressed with zstd if larger than 1MB, by @devdanzin.
 
 ## [0.0.1] - 2024-11-20